requirements.txt
python==3.11.0
aiogram==3.3.0
python-dotenv==1.0.0
bs4==0.0.2
requests==2.32.3
lxml==5.3.0
sqlalchemy==2.0.25
asyncpg==0.29.0
apscheduler==3.11.0
aiosqlite==0.19.0
